Make it possible to specify a sync provider with a parameter
Example:
class MyTimeProvider
{
public:
  time_t timezone;

  time_t get_UTC_from_device(void);

  static time_t ProvideSync(void *p)
  {
    MyTimeProvider *timeprovider = (MyTimeProvider*)p;

    return p->get_UTC_from_device() + p->timezone;
  }

  void MakeMeSyncProvider(void)
  {
    setSyncProvider(ProvideSync, this);
  }
};

It should still be compatible with all old code that doesn't use a callback parameter on the sync provider function. If both a sync-provider-with-parameter and a sync-provider-without-parameter are provided, the one without the parameter is ignored. If you want to cancel a syncprovider-with-parameter, make sure you call setSyncProvider(NULL, NULL) instead of setSyncProvider(NULL).